Rename namespace-registry to conform to go standards and move registry db test
This renames a few things to better conform to go conventions, and it moves the test
Joe put together for testing the client commands closer to the source.
