+++ namespace_registry/client_commands.go
