+++ namespace_registry/registry.go
