+++ namespace_registry/registry_db.go
